requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
facebook-business==18.0.0
apscheduler==3.10.4
sqlalchemy==2.0.44
pytz==2024.2
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.3
